Move PseudoMarcParser implementation
